MatrixChat: Move the event indexing logic into separate modules.
